Updated default config file to better handle 0 RPM
The default config file does not work well with GPUs that force 0 RPM mode. Specifically, a PWM value of 102 is the minimum value that the fans on the ASRock Taichi 5700 XT will respond to. If your GPU's 0 RPM mode requires an even higher PWM value to turn on the fans, please send a pull request with these higher values and the name of the card that needs them.

I've also set the fan curve to be more aggressive in general. Max fan speed still isn't enough to keep hotspot (junction) temps below the throttle point of 110 C, but it will keep it low enough to stop the infamous 5700XT green flashes.